+++ docs/CHANGELOG.md
@@ -3,3 +3,4 @@
 - Fix GitHub actions- 
 - Update H2GIS to 2.2.4-SNAPSHOT
 - Fix maven version
+- H2GIS method to delete database files